extract trunk schema on its own module
